Fix domain boundary regressions
