+++ apps/server/vibesensor/domain/analysis_settings.py
@@ -5,14 +5,68 @@ from __future__ import annotations
 from dataclasses import dataclass
 from typing import ClassVar
 
-from vibesensor.shared.analysis_settings_schema import (
-    ANALYSIS_SETTINGS_BOUNDS,
-    ANALYSIS_SETTINGS_DEFAULTS,
-    ANALYSIS_SETTINGS_NON_NEGATIVE_KEYS,
-    ANALYSIS_SETTINGS_POSITIVE_REQUIRED_KEYS,
+__all__ = ["AnalysisSettingsSnapshot"]
+
+ANALYSIS_SETTINGS_POSITIVE_REQUIRED_KEYS: frozenset[str] = frozenset(
+    {
+        "tire_width_mm",
+        "tire_aspect_pct",
+        "rim_in",
+        "final_drive_ratio",
+        "current_gear_ratio",
+        "wheel_bandwidth_pct",
+        "driveshaft_bandwidth_pct",
+        "engine_bandwidth_pct",
+        "max_band_half_width_pct",
+        "tire_deflection_factor",
+    },
 )
 
-__all__ = ["AnalysisSettingsSnapshot"]
+ANALYSIS_SETTINGS_NON_NEGATIVE_KEYS: frozenset[str] = frozenset(
+    {
+        "speed_uncertainty_pct",
+        "tire_diameter_uncertainty_pct",
+        "final_drive_uncertainty_pct",
+        "gear_uncertainty_pct",
+        "min_abs_band_hz",
+    },
+)
+
+ANALYSIS_SETTINGS_BOUNDS: dict[str, tuple[float, float]] = {
+    "wheel_bandwidth_pct": (0.1, 100.0),
+    "driveshaft_bandwidth_pct": (0.1, 100.0),
+    "engine_bandwidth_pct": (0.1, 100.0),
+    "speed_uncertainty_pct": (0.0, 100.0),
+    "tire_diameter_uncertainty_pct": (0.0, 100.0),
+    "final_drive_uncertainty_pct": (0.0, 100.0),
+    "gear_uncertainty_pct": (0.0, 100.0),
+    "final_drive_ratio": (0.1, 20.0),
+    "current_gear_ratio": (0.1, 20.0),
+    "min_abs_band_hz": (0.0, 500.0),
+    "max_band_half_width_pct": (0.1, 100.0),
+    "tire_width_mm": (100.0, 500.0),
+    "tire_aspect_pct": (10.0, 90.0),
+    "rim_in": (10.0, 30.0),
+    "tire_deflection_factor": (0.85, 1.0),
+}
+
+ANALYSIS_SETTINGS_DEFAULTS: dict[str, float] = {
+    "tire_width_mm": 285.0,
+    "tire_aspect_pct": 30.0,
+    "rim_in": 21.0,
+    "final_drive_ratio": 3.08,
+    "current_gear_ratio": 0.64,
+    "wheel_bandwidth_pct": 5.0,
+    "driveshaft_bandwidth_pct": 4.5,
+    "engine_bandwidth_pct": 5.2,
+    "speed_uncertainty_pct": 1.0,
+    "tire_diameter_uncertainty_pct": 1.0,
+    "final_drive_uncertainty_pct": 0.1,
+    "gear_uncertainty_pct": 0.2,
+    "min_abs_band_hz": 0.2,
+    "max_band_half_width_pct": 6.0,
+    "tire_deflection_factor": 0.97,
+}
 
 
 @dataclass(frozen=True, slots=True)
